+++ src/main/java/de/samply/notification/OperationType.java
@@ -1,4 +1,5 @@
 package de.samply.notification;
 
 public enum OperationType {
+    GENERATE_TOKEN
 }