+++ CHANGELOG.md
@@ -72,3 +72,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Accept, Reject and request changes in project results
 - Notification Service
 - Frontend DTO
+- Email as Mime message